Remove CSS style attribute from `dicono-di-me.html` because it did not improve scrolling on iOS.
Fixed mobile version of `dicono-di-me.html` as the previous commit obviously broke it.
Figured out why there was the JS WTF in `mobile_about_us.js`. When setting `className = ''`, apparently the array of extracted elements with `getElementsByClassName was getting auto updated; so it was shrinking one element when removing the className attribute from one of its elements (!!)